fix (skeleton-typescript-aspnetcore): change VS Task Runner binding to run the bundle task after building
In order to see changes made in the project source, the bundle task must be run which updates the files in the dist directory. By binding this gulp task to VS build, it will ensure the changes will be seen when the project is run.  Otherwise, the user must manually run the bundle task in order to see the changes.